backend/requirements.txt
# requirements.txt for backend dependencies
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
